tests/wpt/tests/css/css-view-transitions/parsing/view-transition-class-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS View Transitions Test: view-transition-class with invalid values</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-1/">
<meta name="assert" content="view-transition-class does not support invalid values">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("view-transition-class", "default"); // `default` isn't allowed by the `<custom-ident>` syntax.
test_invalid_value("view-transition-class", "foo none");
test_invalid_value("view-transition-class", "#fff");
test_invalid_value("view-transition-class", "12px");
test_invalid_value("view-transition-class", "foo 12px");
test_invalid_value("view-transition-class", "12em");
test_invalid_value("view-transition-class", "12%");
</script>
</body>
</html>
